Pool2D implementation for CPU backend with mkldnn
Summary:
Implements pooling with the CPU backend with MKL-DNN. Supports max pooling and average pooling (with and without padding).

Unless the forward pass is done in inference mode, mkldnn requires that both workspace memory and a pooling primitive descriptor used in the forward pass be passed to backwards primitive descriptors so as to better optimize gradient computation. Memory descriptors for both of these are captured/saved on the stack for gradFunc, since they can't be reconstructed without the forward descriptor.

Reviewed By: vineelpratap

Differential Revision: D13579084

fbshipit-source-id: 4a7521644f11352f710b2d27eeddd90127e93d95

+++ tests/autograd/AutogradTest.cpp
@@ -17,9 +17,9 @@
 #include <functional>
 #include <stdexcept>
 
+#include <gtest/gtest.h>
 #include "flashlight/autograd/autograd.h"
 #include "flashlight/common/common.h"
-#include <gtest/gtest.h>
 
 using namespace fl;
 
@@ -32,7 +32,8 @@ bool jacobianTestImpl(
     float precision = 1E-5,
     float perturbation = 1E-4) {
   auto fwd_jacobian =
-      af::array(func(input).elements(), input.elements(), af::dtype::f64);
+      af::array(func(input).elements(), input.elements(), af::dtype::f32);
+
   for (int i = 0; i < input.elements(); ++i) {
     af::array orig = input.array()(i);
     input.array()(i) = orig - perturbation;
@@ -46,7 +47,7 @@ bool jacobianTestImpl(
   }
 
   auto bwd_jacobian =
-      af::array(func(input).elements(), input.elements(), af::dtype::f64);
+      af::array(func(input).elements(), input.elements(), af::dtype::f32);
   auto dout =
       Variable(af::constant(0, func(input).dims(), input.type()), false);
   for (int i = 0; i < dout.elements(); ++i) {
@@ -560,7 +561,7 @@ TEST(AutogradTest, BinaryCrossEntropy) {
 TEST(AutogradTest, CrossEntropy) {
   auto in = Variable(af::randu(7, 10, 1, 1, af::dtype::f64), true);
   af::dim4 dims(10, 1, 1, 1);
-  auto y = Variable(af::constant(1.0, dims, af::dtype::f32), false);
+  auto y = Variable(af::constant(1.0, dims, af::dtype::f64), false);
 
   auto func_crossent_mean = [&](Variable& input) {
     return categoricalCrossEntropy(input, y);
